Add information about partial events & ctors
