+++ docs/csharp/language-reference/keywords/snippets/events.cs
@@ -1,0 +1,45 @@
+namespace Events;   
+   
+//<EventExample>
+public class SampleEventArgs
+{
+    public SampleEventArgs(string text) { Text = text; }
+    public string Text { get; } // readonly
+}
+
+public class Publisher
+{
+    // Declare the delegate (if using non-generic pattern).
+    public delegate void SampleEventHandler(object sender, SampleEventArgs e);
+
+    // Declare the event.
+    public event SampleEventHandler SampleEvent;
+
+    // Wrap the event in a protected virtual method
+    // to enable derived classes to raise the event.
+    protected virtual void RaiseSampleEvent()
+    {
+        // Raise the event in a thread-safe manner using the ?. operator.
+        SampleEvent?.Invoke(this, new SampleEventArgs("Hello"));
+    }
+}
+//</EventExample>
+
+interface IDrawingObject
+{
+    event EventHandler OnDraw;
+}
+
+//<AddHandler>
+class Events : IDrawingObject
+{
+    event EventHandler PreDrawEvent;
+
+    event EventHandler IDrawingObject.OnDraw
+    {
+        add => PreDrawEvent += value;
+        remove => PreDrawEvent -= value;
+    }
+}
+//</AddHandler>
+
