+++ docs/csharp/specification/toc.yml
@@ -201,6 +201,8 @@ items:
       href: ../../../_csharplang/proposals/csharp-9.0/extending-partial-methods.md
     - name: Partial properties
       href: ../../../_csharplang/proposals/csharp-13.0/partial-properties.md
+    - name: Partial events and constructors
+      href: ../../../_csharplang/proposals/partial-events-constructors.md
     - name: Field backed properties (preview)
       href: ../../../_csharplang/proposals/field-keyword.md
     - name: Covariant return types
