+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csrefKeywordsContextual/CS/csrefKeywordsContextual.cs
@@ -78,21 +78,4 @@ namespace csrefKeywordsContextual
         }
 
     }
-    interface IDrawingObject
-    {
-        event EventHandler OnDraw;
-    }
-
-    //<Snippet15>
-    class Events : IDrawingObject
-    {
-        event EventHandler PreDrawEvent;
-
-        event EventHandler IDrawingObject.OnDraw
-        {
-            add => PreDrawEvent += value;
-            remove => PreDrawEvent -= value;
-        }
-    }
-    //</Snippet15>
 }
