+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csrefKeywordsModifiers/CS/csrefKeywordsModifiers.cs
@@ -214,31 +214,6 @@ namespace csrefKeywordsModifiers
     // Output: My local constant = 707
     //</snippet6>
 
-    //<snippet7>
-    public class SampleEventArgs
-    {
-        public SampleEventArgs(string text) { Text = text; }
-        public string Text { get; } // readonly
-    }
-
-    public class Publisher
-    {
-        // Declare the delegate (if using non-generic pattern).
-        public delegate void SampleEventHandler(object sender, SampleEventArgs e);
-
-        // Declare the event.
-        public event SampleEventHandler SampleEvent;
-
-        // Wrap the event in a protected virtual method
-        // to enable derived classes to raise the event.
-        protected virtual void RaiseSampleEvent()
-        {
-            // Raise the event in a thread-safe manner using the ?. operator.
-            SampleEvent?.Invoke(this, new SampleEventArgs("Hello"));
-        }
-    }
-    //</snippet7>
-
     //<snippet8>
     //using System.Runtime.InteropServices;
     class ExternTest
